requirements.txt
colorama
python-guerrillamail